Added a new plantuml-preview-theme variable
When the plantuml-preview-theme variable is set the string sent to
the preview buffer command is modified and the !theme directive is
inserted after @startuml.

This is useful when you don't want to modify the plantuml source to make
the preview suite your theme in Emacs.

(cherry picked from commit f30bea6a5c9b53712e4608247c79d62d62c53136)